moved getPricesToUpdate method from JSON model to service; renamed package
