+++ src/main/java/org/galatea/jingyang/financepriceservice/domain/OneDayPrice.java
@@ -1,4 +1,4 @@
-package org.galatea.jingyang.finance_price_service.domain;
+package org.galatea.jingyang.financepriceservice.domain;
 
 import java.io.Serializable;
 import javax.persistence.Column;
@@ -10,7 +10,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.galatea.jingyang.finance_price_service.domain.OneDayPrice.OneDayPriceId;
+import org.galatea.jingyang.financepriceservice.domain.OneDayPrice.OneDayPriceId;
 
 @AllArgsConstructor
 @NoArgsConstructor
