+++ src/main/java/org/galatea/jingyang/financepriceservice/domain/PricesSet.java
@@ -1,4 +1,4 @@
-package org.galatea.jingyang.finance_price_service.domain;
+package org.galatea.jingyang.financepriceservice.domain;
 
 import java.util.ArrayList;
 import lombok.Builder;
