+++ src/main/java/org/galatea/jingyang/financepriceservice/domain/modelresponse/AlphaVantageAPIKeywords.java
@@ -1,4 +1,4 @@
-package org.galatea.jingyang.finance_price_service.domain.alpha_vantage_objects;
+package org.galatea.jingyang.financepriceservice.domain.modelresponse;
 
 import lombok.experimental.UtilityClass;
