+++ src/main/java/org/galatea/jingyang/financepriceservice/domain/rspy/ClosedDatesRepository.java
@@ -1,6 +1,6 @@
-package org.galatea.jingyang.finance_price_service.domain.rspy;
+package org.galatea.jingyang.financepriceservice.domain.rspy;
 
-import org.galatea.jingyang.finance_price_service.domain.ClosedDate;
+import org.galatea.jingyang.financepriceservice.domain.ClosedDate;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ClosedDatesRepository extends CrudRepository<ClosedDate, Long> {
